refactor: move recorder-history transforms into history.py
``extract_numeric_values``, ``extract_timestamped_numeric_values`` and
``resample_history`` are pure functions over recorder output — no
``hass``, no I/O, no idea who asked. They sat at module level in
coordinator.py, which made the coordinator the import target for
anything that wanted them: the websocket preview reached back into it
for ``resample_history`` through a deferred import specifically to dodge
the cycle.

Give them their own module. The rules they encode are about history
data (what counts as numeric, which binary states map to 1/0, how a step
signal is resampled onto a time axis), and keeping them apart from the
orchestration that calls them means the upcoming resolver can depend on
history without depending on the coordinator.

Pure move: no behavior change, tests only retarget their imports.
