+++ custom_components/geekmagic/websocket.py
@@ -430,8 +430,7 @@ def _resample_history(history_states: list, start: datetime, end: datetime) -> l
     Returns:
         Time-faithful resampled values for the chart sparkline.
     """
-    # Import the shared function from coordinator
-    from .coordinator import resample_history
+    from .history import resample_history
 
     return resample_history(history_states, start, end)
